requirements.txt
fastapi>=0.104.0

uvicorn[standard]>=0.24.0

prometheus-fastapi-instrumentator>=6.1.0

pydantic>=2.0.0

joblib>=1.3.0

pandas>=2.0.0

python-dotenv>=1.0.0

python-jose>=3.3.0

redis>=5.0.0

scikit-learn>=1.3.0

python-jose[cryptography]>=3.3.0